backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.0
python-dotenv==1.0.0
openai==1.3.5
motor==3.3.2
pymongo==4.6.0
httpx==0.25.2
requests==2.31.0
python-multipart==0.0.6
chromadb==0.4.18
numpy==1.24.3
